Refactoring of datasets
Renamed table column to the python's convention format
