+++ backend/dataall/modules/datasets/api/__init__.py
@@ -1,6 +1,6 @@
 """The GraphQL schema of datasets and related functionality"""
 from dataall.modules.datasets.api import (
-    DatasetTableColumn
+    table_column
 )
 
-__all__ = ["DatasetTableColumn"]
+__all__ = ["table_column"]
