+++ backend/dataall/modules/datasets/api/table_column/queries.py
@@ -1,5 +1,5 @@
 from dataall.api import gql
-from dataall.modules.datasets.api.DatasetTableColumn.resolvers import list_table_columns
+from dataall.modules.datasets.api.table_column.resolvers import list_table_columns
 
 listDatasetTableColumns = gql.QueryField(
     name='listDatasetTableColumns',
